Removed UIKit ScreenSaver implementation: Simplified ScreenSaverManager: ❌ Removed oiverlay window gesture recognizers
