+++ openHAB/ScreenSaver/ScreenSaverManager.swift
@@ -30,8 +30,6 @@ final class ScreenSaverManager: NSObject {
 
     private weak var window: UIWindow?
 
-    private var saverView: ScreenSaverView?
-
     private var overlayWindow: UIWindow?
 
     private var previousBrightness: CGFloat?
@@ -52,7 +50,7 @@ final class ScreenSaverManager: NSObject {
 
     public func updateConfiguration(_ newConfiguration: ScreenSaverConfiguration) {
         configuration = newConfiguration
-        if saverView != nil {
+        if overlayWindow != nil {
             dismissSaverIfNeeded()
         }
 
@@ -96,7 +94,7 @@ final class ScreenSaverManager: NSObject {
 
     private func showSaver() {
         guard configuration.isEnabled else { return }
-        guard saverView == nil, let baseWindow = window else { return }
+        guard overlayWindow == nil, let baseWindow = window else { return }
         logger.debug("Presenting screen saver (overlay window)")
 
         let overlay: UIWindow
@@ -116,36 +114,7 @@ final class ScreenSaverManager: NSObject {
 
         hostVC.setNeedsStatusBarAppearanceUpdate()
 
-        let swiftUIView: any View
-        if #available(iOS 16.0, *) {
-            swiftUIView = ScreenSaverSwiftUIView(configuration: configuration)
-        } else {
-            // Fallback to UIKit version for older iOS
-            let saver = ScreenSaverView(configuration: configuration)
-            saver.translatesAutoresizingMaskIntoConstraints = false
-            hostVC.view.addSubview(saver)
-            NSLayoutConstraint.activate([
-                saver.leadingAnchor.constraint(equalTo: hostVC.view.leadingAnchor),
-                saver.trailingAnchor.constraint(equalTo: hostVC.view.trailingAnchor),
-                saver.topAnchor.constraint(equalTo: hostVC.view.topAnchor),
-                saver.bottomAnchor.constraint(equalTo: hostVC.view.bottomAnchor)
-            ])
-
-            attachGestureRecognizers(to: overlay)
-
-            saver.alpha = 0
-            UIView.animate(withDuration: 0.3) {
-                saver.alpha = 1.0
-            } completion: { _ in
-                saver.startAnimation()
-            }
-
-            saverView = saver
-            overlayWindow = overlay
-            applyDimming()
-            return
-        }
-
+        let swiftUIView = ScreenSaverView(configuration: configuration)
         let hostingController = UIHostingController(rootView: AnyView(swiftUIView))
         hostingController.view.backgroundColor = .clear
         hostingController.view.translatesAutoresizingMaskIntoConstraints = false
@@ -161,16 +130,11 @@ final class ScreenSaverManager: NSObject {
             hostingController.view.bottomAnchor.constraint(equalTo: hostVC.view.bottomAnchor)
         ])
 
-        // wake up if the user taps anywhere
-        attachGestureRecognizers(to: overlay)
-
         hostingController.view.alpha = 0
         UIView.animate(withDuration: 0.3) {
             hostingController.view.alpha = 1.0
         }
 
-        // Store reference for SwiftUI case (nil for UIKit fallback which handles its own storage)
-        saverView = nil
         overlayWindow = overlay
         applyDimming()
     }
@@ -178,9 +142,6 @@ final class ScreenSaverManager: NSObject {
     private func dismissSaverIfNeeded() {
         guard overlayWindow != nil else { return }
         logger.debug("Dismissing screen saver")
-        if let saver = saverView {
-            saver.stopAnimation()
-        }
         if configuration.enablesAutoDimming {
             if configuration.restoresBrightness {
                 restoreBrightnessIfNeeded()
@@ -197,7 +158,6 @@ final class ScreenSaverManager: NSObject {
             self.overlayWindow?.isHidden = true
             self.overlayWindow = nil
         }
-        saverView = nil
     }
 
     private func applyDimming() {
